+++ kifmm/src/fftw.rs
@@ -9,6 +9,7 @@
 //! * Optionally parallel `r2c` and `c2r` which share a plan to compute batched transforms of multiple input data sets.
 //!
 mod c2c;
+mod dft;
 mod helpers;
 mod r2c;
 pub mod types;
